project1/requirements.txt
requests
RPi.GPIO
gpiozero